Fix Base ETH & Other Native Assets (#51)
Base ETH Token that here was incorrect:

https://basescan.org/address/0x575a2bB6F8C31f8e60264531d5e300e6EC505F29

ETH is a native asset on base so, this should do it.

Second commit adds native assets to other chains where it was missing.


This also was the cause of a Not Found error in the near-cow-agent for
the token pair:

```
{
  "sellToken": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
  "buyToken": "0x575a2bB6F8C31f8e60264531d5e300e6EC505F29",
}
```

Not Found was because there was no trading route found.