Split Token Map (#53)
It was impossible to tell the chainId with them all in one file. Now
each chain has its own dedicated file. This is still far from the best
solution, but its better. Makes updates more searchable. Still need to
update the generation script and align with this Dune Query.

https://dune.com/queries/4362544

Note that generation script would override all manual changes...

This should be backwards compatible with the old token map file (since
we didn't remove it). However, it should be considered deprecated.